requirements.txt
argon2-cffi==19.2.0
asgiref==3.2.5
bcrypt==3.1.7
certifi==2019.11.28
cffi==1.14.0
dj-database-url==0.5.0
Django==3.0.3
django-bootstrap3==12.0.3
django-braces==1.14.0
django-crispy-forms==1.9.0
django-filter==2.2.0
django-heroku==0.3.1
gunicorn==20.0.4
misaka==2.1.1
Pillow==7.1.1
psycopg2==2.8.5
pycparser==2.20
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
whitenoise==5.0.1
wincertstore==0.2
